Reset data input elements when no valid selection present
This change makes the UI behaviour more consistent - if the user
changes the input text but does not select a valid item from the
suggestion list all input elements (hidden & visible) are reset to
empty. This gives the user a visual prompt that no valid entry was made
and allows the app to handle that in its own validation.

Our solution to this issue
https://github.com/crowdint/rails3-jquery-autocomplete/issues/83